+++ R/helpfunctions.R
@@ -55,3 +55,4 @@ get_grouped_data <- function(x) {
 
   grps
 }
+